refactor: moved download component to shared, so it can be used in admin
